Implement unknown field support
This PR implements unknown fields via a somewhat generic mechanism:
every message can now have an optional "cold struct" which contains
rarely-used fields, allowing these fields to not contribute to memory
usage by default. The unknown
fields live in an array of zero-copy byte slices in this struct.

Currently, unknown fields are always preserved, but it would be very
easy to add a flag to `typeParser` that instructs it to discard all
unknown fields.

The new cold pointer does not take up extra space in `message`; to
achieve this, the old `ty` field is now a 32-bit offset, and the
remaining 32 bits are now an index into a slice of cold objects.