Implement extension support
This PR adds support for extensions, which are pre-computed ahead of
time for any given message, instead of being looked up at runtime
mid-parse.

Extensions are treated as normal fields almost everywhere in the parser,
and some type data structures have been updated to use an internal list
of `FieldDescriptor`s that includes the extensions in the order we
happened to get them in, so that we can look up field descriptors
quickly for our implementation of the reflection API.

The main contribution of this PR is the concept of "cold fields", which
are fields whose storage is behind a lazily-allocated pointer instead of
in the standard struct (this is also where the unknown fields live, from
a previous PR). The compiler can arbitrarily decide to make any field
cold, and a generic mechanism is provided via PGO for marking fields as
hot or cold. By default, extensions are cold, and non-extensions are
hot.