Implement support for scalar-valued maps
This PR adds all map types, except for those of the form `map<K,
Message>`, which will require some more complex machinery.

In addition, this PR adds a swisstable implementation to replace
`table.Table` by making it generic. The change is not especially big,
since `table.Table` already used the same probing strategy as
swisstable. This change also makes extensive use of `//fastpb:stencil`
to optimize table lookups.

`map<bool, V>` is given a special implementation. Instead of being a
map, it is implemented as a pair of `optional` fields. This optimization
probably isn't especially necessary because `map<bool, V>` is quite
rare, but not doing it also felt wrong to me. `map<bool, bool>` is just
four bits.

Scalar map parsing is 4-5x faster than it is for gencode.

```
                                   |       time |    thoughput |     memory |    allocations
./maps/big/v32xz32_small/fastpb    |  62.927 μs | 227.820 MB/s |  24.670 kB | 29.000  allocs
 ''                     /amortize  |  61.353 μs | 233.660 MB/s |  17.018 kB | 24.000  allocs
 ''                     /gencode   | 235.682 μs |  60.830 MB/s |  28.684 kB |  6.129 kallocs
 ''                     /dynamicpb | 241.654 μs |  59.320 MB/s |  19.396 kB |  2.041 kallocs
```