Optimize overhead for small messages
This change makes it so that parsing a small message is actually faster than
using the gencode. Previously, various unnecessary overheads (such as
construction of memory block types in reflection, redundant calls to
`Arena.KeepAlive`) made parsing small (~16 byte) messages slower in our
implementation.

This PR fixes that, and adds a small benchmark specifically for
measuring this.

```
                         |       time |    thoughput |     memory |    allocations
./shallow/#00/fastpb     | 304.900 ns |  13.120 MB/s | 878.000  B |  4.000  allocs
 ''          /amortize   |  88.610 ns |  45.140 MB/s |    0.000 B |   0.000 allocs
 ''          /gencode    | 157.500 ns |  25.400 MB/s | 168.000  B |  3.000  allocs
 ''          /dynamicpb  | 686.900 ns |   5.820 MB/s |   1.128 kB | 16.000  allocs
```