Reorganize the `field_*.go` files
This changes `stencil.go` to put all of the generated stencils in one
file, reducing the amount of clutter in the root directory. It also
splits some of the `field_*.go` files, and adds some documentation
around how fields are implemented and what all the files actually have
in them.