Implement message-valued maps
This adds support for message-valued maps.

Message-valued maps are quite tricky, because we need to determine the
key of an entry before we call the message parsing routine, which the
parser thunk "calls" by calling `p1.message()` and returning. To solve
this, we assume that most entries are laid out with the key first, and
optimize the parser function for that case. However, all of the
pathological cases are also handled (with tests to match).

As usual, a benchmark is included.

```
                             |       time |   throughput |     memory |    allocations
./maps/big/v32xmsg/fastpb    | 263.026 μs | 154.940 MB/s |   1.139 MB | 18.000  allocs
./''''/'''/'''''''/amortize  | 181.612 μs | 224.400 MB/s |  41.219 kB |  1.000  allocs
./maps/big/v32xmsg/gencode   | 536.191 μs |  76.000 MB/s | 963.286 kB |  6.156 kallocs
./''''/'''/'''''''/dynamicpb |   1.166 ms |  34.940 MB/s |   1.859 MB | 18.683 kallocs
```