Make all `protoreflect.List` implementations have pointer receiver
This ensures that accessing a `List` through reflection does not trip
an allocation.

This required significant refactoring of the `unsafe2` package, because
we needed somewhat more subtle pointer management to avoid needing a
pointer back to the message root in all cases. As a result:
- `fastpb.zc` is now in its own package and callsed `zc.Range`.
- `arena.Slice` is now in its own package and called `slice.Slice`;
there are also helpers for using the sign bit of the pointer value as a
tag.
- Layout helpers got moved into `unsafe2/layout`, and now we have a
proper `layout.Layout` type. Also we can now get rid of all the annoying
size/align constants.

The new `unsafe2/support` package prevents `fastpb` from building on any
architecture that we do not explicitly support, which for now is only
x86_64 and aarch64.