Fix subtle non-determinsm in `internal/tools/stencil`
The package that was selected for the generated stencil file could turn
out to be one of `pkg` or `pkg_test` when generating a
`stencils_test.go` file. This was due to an incorrect assumption that
all `_test.go` files would be in `pkg_test`, which lead to a race
condition.

This PR fixes that race condition by deciding on the package to use
before parsing the files.