Pull the field thunks into their own package, `internal/tdp/thunks`
This is a followup that pulls more internal details out of the
root package.

In order to make this work and break circular dependencies, we need to
add two callbacks that are initialized dynamically, for constructing
`protoreflect.Value`s within the `thunks` package, which is a dependency
of the package that defines the actual `protoreflect.Message` and
`protoreflect.Type` types.