pick f9ab6eff Prepare for open-sourcing pick dfedcca4 Add CI checks pick 168a29dd Unbreak CI (#1) pick 81133947 Implement oneof support (#3) pick a57d1524 Split `field.go` according to archetype classes (#4) pick 087c4925 Implement `Message.WhichOneof` (#5) pick e1251af8 Implement unknown field support (#6) pick 31bd93d5 Add more CI targets (#7) pick 30622193 Reorganize the tests, fix several bugs with repeated fields (#8) pick bdaf8e0e Fix some issues with benchmarks (#9) pick bdbdf8d2 Implement extension support (#10) pick adf067f3 Add a script for generating a pretty benchmark report for putting in commit messages (#12) pick df40494b Implement support for scalar-valued maps (#11) Fix a memory safety bug in `vm.parseVarint()`
Although I recently added a check to `P1.Ptr()` to ensure that we do not
create one-past-the-end pointers, there is one place, in `parseVarint`,
where we increment a pointer *before* doing a necessary bounds check.

This PR fixes that, by carefully sequencing loads and bounds checks. It
also adds a `GC()` call in debug mode to catch potential mistakes. The
new version of `parseVarint` has been benchmarked and performance is not
statistically different.