Add `required` checking and other compiler options
The main thing added is efficient `required` checking. We perform an
inter-message analysis to determine if a message definitely has no
transitive required fields, allowing us to skip the check altogether for
basically every message. The analysis consists of sorting the message
graph with Tarjan's algorithm, and marking each scc as containing
required fields if any message in it has a required field, or if any of
of the sccs it depends on contain required fields. Then, a field only
needs to be
recursed into to check for required fields if it has a message type
whose scc was marked by the above analysis.

This also adds an option for discarding unknown fields, and makes
aliasing the input buffer opt-in with an option. Unsafe operations for
speeding up reflection have been moved to the `protoreflect2` package.