Add special cases for the first 15 field numbers
The first 15 field numbers are very common, and quite popular for hot
fields. This change adds a special fast path for when we encounter a
one-byte varint at the start of a record. If we do, we check a new
special lookup table which is populated with the offsets of the first 15
field numbers' parsers. Parsers are only present here if they are very
close to the parser header, which is usually the case. If there isn't an
entry in the table, we still skip all of the "large" varint decoding
code and jump straight into the field parser walker loop.

This seems to result in noticable ~5-10% speedups, such as in the
descriptor.yaml benchmark.