Implement support for groups
This PR adds support for one final Protobuf feature: groups. Although
groups are deprecared in `proto2` and not present in `proto3`, there is
nothing stopping users from needing to parse them, particularly since
they are available in editions mode.

Groups are implemented by retaining the end-group tag in `vm.P1`, and
checking for it whenever parsing a regular tag fails enough times. An
alternative strategy would be to generate a unique `tdp.TypeParser` for
each group-typed field, which includes the group end tag among its
fields. This would eliminate the cost of checking for group termination
for messages. However, that would be fairly tricky to implement, since
popping a message inside of a thunk probably won't work right.

There were several places where we had to adjust the "message tail call"
optimization to account for groups, since groups can't be tail-called
(either as callee or caller).