Remove a confounding variable from the allocator in benchmarks
Before this change, the memory passed into unmarshal benchmarks may or
may not trigger the memory conditioning function in the VM to ensure
that it can read 9 bytes past the end. Whether it actually did so seemed
to be somewhat random.

This change:
1. Fixes the conditioning function to ensure there are nine bytes beyond
the length of the slice, either between length and capacity or between
the end of the length and the next page boundary.
2. Calls the conditioning function on each benchmark specimen.

As part of all this I was trying to whack other seemingly wasteful
annotations, and I found that the `vm.Options` used to collect
unmarshaling options gets spilled unnecessarily because it escapes into
a func. To avoid this, I've made it impossible for users to construct
their own `UnmarshalOption`s (previously they could using
`reflect.MakeFunc`), and then hide the `*vm.Options` passed into the
option func from escape analysis using `unsafe2.NoEscape`.

Unfortunately, I had to use a slightly weird representation for
`UnmarshalOption` due to a Go escape analysis bug I discovered:
https://github.com/golang/go/issues/74356.