Optimize maps some more
This PR fixes a few random codegen issues with the map parsers:

1. `varintItem[...].parse` did not like inlining, which was causing
heavy register motion in parsing functions that used it. It's now
specialized so that even if it doesn't inline, it doesn't require a
generic dictionary parameter.
2. `zc.Range.Bytes` was not inlining in a critical spot. Added a
workaround.

This seems to have gained us some percentage points in map parsing. In
particular, the v32xz32 map benchmark is now competitive with vtproto.