Add new optimizations around UTF-8 and packed field counting
This change optimizes a couple of the "sign bit counting" routines to
use a common pattern in vectorized algorithms, which is a manual kind of
unroll-and-jam optimizations that fissions off the last iteration, so
that you don't need to do the mask-off-extra-bits operation each
iteration.

See https://mcyoung.xyz/2023/11/27/simd-base64/#unroll-and-jam-revisited
for an example of this technique.

With this change in, we are now competitive with vtproto in some
string-heavy benchmarks (note that, again, vtproto cheats by skipping
UTF-8 validation completely).