Remove uses of `protoreflect.ValueOf` in accessors
Due to https://github.com/golang/go/issues/74364,
`protoreflect.ValueOf()` escapes its `any` argument if its return value
escapes, meaning that *any* call to `protoreflect.ValueOf()` with a
scalar will allocate. This is an unacceptable performance hit.

This PR avoids the problem by replacing all calls to `ValueOf` with a
typed version, or in the case of generic functions, a call to
`protoreflect2.ValueOfScalar()`, which is written to avoid this
performance degradation.