Make sure that everything builds with `-race` (#10)
There is no concurrency in `hyperpb`, but users might decide to build
with `-race`, so we should make sure that we can build (and pass) with
the race detector on; this is similar to how tests are run with
`-d=checkptr`.

However, building with `-race` fails the nosplit check. This is easily
fixed by disabling race instrumentation in one function (which requires
generalizing the nosplit handling code in `hyperstencil`).