Temporarily make hyperpb build without optimizations turned on (#31)
This PR mitigates #30 at the cost of some performance.

Unfortunately, some tools disable optimizations, despite the fact that
Go is not especially aggressive at optimizing code, nor does Go actually
support doing this (seeing as the nosplit check is an
optimization-dependent diagnostic). This should be fixed in upstreams
inappropriately disabling optimizations; once that is fixed, we will
stop allowing users to build without them.

In this PR, we comment out `go:nosplit` directives to achieve this goal,
and add a CI check to prevent backsliding.